# ui/termstatus: Optimize Truncate