@greatroar 2023-04-14
Previously they were considered wide, now they're narrow. I initially added a test, then figured that the behavior doesn't matter because we shouldn't output invalid UTF-8 anyway.